Add more general ∫G function and spheroid test
∫G function generalizes the quadrature a bit more so that I can chose to use different quadrature levels for the source and the reflected source. This is a superficial change, but it feels a bit more consistent to me.

The disturbance potential ϕ still needs to have a branch because I have to use `p` and `reflect(p)` in ∫G for the source and its reflection. I don't see a way around this, unfortunately.

I've also updated the spheroid geometry to better calculate the spacings and added a surface pirecing