Multi-threading
Added multi-threading to functions that broadcast or reduce over panels. New multi-threaded functions `influence` to create A, `steady_drag` to give the integrated pressure coefficient and `added_mass` to create the full added mass matrix (although there is a bug in the zz component of the circle example).

Also multi-threaded computing φ(x) and the data generation for `chebinterp` but unfortunately, the precompile is always in serial. I'll move it elsewhere.