Add dz to ∫kelvin and make wavelike even in α
I've switched to using the 2x2 quadrature for kelvin is a Froude-number scaled vertical distance is less than a threshold dz (default 10). This increases the accuracy of the quadrature near z=0 when using bigger panels.

I've also enforced the fact that nearfield is an even function in α. This has basically no effect, but now the y derivative on y=0 is exactly zero.