Update param_props
Sampling Guass points on the tangent-plane can put some of them above the waterline (which I've kept only a warning). I've made `tangentplane=true` an argument, and it subsamples the surface itself otherwise.

The surface piercing spheroid results improved with `tangentplane=false` and the new `ϕ=∫kelvin` and I've updated the figure.