Area fix
The first order areas (that work so well for basic shapes) seem to be very unstable for NURBS. I've replaces it with the area of two triangles touching the corners. I also removed the tangentplane flag.

Also split the strips based on the longer of the top.bottom edges. And I added logic to turn off the viz vectors.