Finally get curved panels working?
Got the 8-panel sphere working with curved-panels. The boundary condition is applied at the _centroid_ of the panel and the area weights are used in the ∫G 2x2 quadrature. I no longer have to cheat the parametric center _and_ the result is x/y/z symmetric! Much better!

I've put a flag on the cubature calls. When not flagged, the 2x2 gauss points are used to get the area and centroid. We don't need it for the sphere, but I've used `cubature=true` to test we get really good area, centroid and symmetry.