Imrpoved thin-ship Wigley example
There is no need to solve for q in thin-ship, it's proportional to n_x. I've updated the test to use this, still placing the panels on the centerplane. However, for finite thickness, you can still place panels on the hull using the thin-ship estimate simply for the density. I've updated the "thin" Wigley to use this more general approach.