switch real-line integral to quadgk
I found that the spectral error level in contour depended strongly on z. The 16 point GL quadrature was overdoing it for large z and not sufficient for small z.
I've switched it over to quadgk. This does allocate a little (and I'll need to check if that's a problem later) but it immediately solved this problem and let's me get rid of some of the hacky stuff I was doing before.