Barnes-Hut evaluation using ImplicitBVH
Evaluating the potential (or it's derivatives) at a target is O(N), leading to O(M*N) scaling for M targets. I'm setting up a Barnes-Hut method to bring that down to O(log N).

ImplicitBVH creates a bounded volume hierarchy which we can apply to the panels. Then we accumulate the panel properties we need (dA, n, x, and q strength) onto the nodes and use those when well outside a volume.

Currently, I have the generic acculumate! and evaluate functions working. I need to add the panel specific logic and tests. Then I want to switch to an matrix free solver to avoid the O(N^2) construction cost entirely.