Add PanelSystem type, and refactor BarnesHut code
Added PanelSystem <: AbstractPanelSystem to give a uniform interface for solving and measuring. Now we have Φ(x,sys) which specializes on the system type, and nearly everything else is uniform.

I renamed the matrix-free solver to GMRESsolver!(sys) and it works on either type. It's in a new file gmres.jl with it's own tests.

I added a nice surface_integral function, which is called for added_mass and steady_force.

Finally, I split the BarnesHutCore functions into their own file which (potentially) could be moved into ImplicitBVH in the future.

kelvin and reflected potentials are untested with BarnesHut.